requirements.txt
acme==1.21.0
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
bcrypt==5.0.0
beniget==0.4.1
blinker==1.9.0
certbot==1.21.0
certbot-nginx==1.21.0
certifi==2025.11.12
cffi==2.0.0
chardet==4.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.4
ConfigArgParse==1.5.3
configobj==5.0.6
contourpy==1.3.2
cryptography==3.4.8
cycler==0.12.1
dbus-python==1.2.18
decorator==4.4.2
distro==1.7.0
distro-info==1.1+ubuntu0.2
exceptiongroup==1.3.0
ffmpeg==1.4
Flask==3.1.2
fonttools==4.60.1
frozenlist==1.8.0
gast==0.5.2
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
idna==3.11
importlib-metadata==4.6.4
invoke==2.2.1
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.6
joblib==1.5.2
josepy==1.10.0
Js2Py==0.74
keyring==23.5.0
kiwisolver==1.4.9
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
magic-filter==1.0.12
MarkupSafe==3.0.3
matplotlib==3.10.7
more-itertools==8.10.0
multidict==6.7.0
noisereduce==3.0.3
numpy==2.2.6
oauthlib==3.2.0
olefile==0.46
outcome==1.3.0.post0
packaging==25.0
paramiko==4.0.0
parsedatetime==2.6
pillow==12.0.0
ply==3.11
propcache==0.4.1
py3xui==0.5.3
pyaes==1.6.1
pyasn1==0.6.1
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.10
pydantic_core==2.33.2
pydub==0.25.1
PyGObject==3.42.1
PyICU==2.8.1
pyjsparser==2.7.1
PyJWT==2.3.0
PyNaCl==1.6.1
pyOpenSSL==21.0.0
pyparsing==3.2.5
pyRFC3339==1.1
Pyrogram==2.0.106
PySocks==1.7.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pythran==0.10.0
pytz==2022.1
PyYAML==5.4.1
requests==2.32.5
requests-toolbelt==0.9.1
rsa==4.9.1
scipy==1.15.3
SecretStorage==3.3.1
selenium==4.38.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
ssh-import-id==5.11
systemd-python==234
Telethon==1.42.0
TgCrypto==1.2.5
tqdm==4.67.1
trio==0.32.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
undetected-chromedriver==3.5.5
urllib3==1.26.20
wadllib==1.3.6
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.3.1
xkit==0.0.0
yarl==1.22.0
yt-dlp==2025.11.12
zipp==1.0.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.1.0
zope.interface==5.4.0
